Day 12: Inheritance
You are given two classes, Person and Student, where Person is the base class and Student is the derived class. Completed code for Person and a declaration for Student are provided for you in the editor. Observe that Studentinherits all the properties of Person.
Complete the Student class by writing the following:
A Student class constructor, which has  parameters:
A string, .
A string, .
An integer, .
An integer array (or vector) of test scores, .
A char calculate() method that calculates a Student object’s average and returns the grade character representative of their calculated average:
Sample Input:
Heraldo Memelli 8135627
2
100 80
Sample Output:
 Name: Memelli, Heraldo
 ID: 8135627
 Grade: O